--- packages/client/src/App.tsx
@@ -1,31 +1,0 @@
-import { useEffect } from 'react'
-import './App.css'
-import { createBrowserRouter, RouterProvider } from 'react-router-dom'
-
-const router = createBrowserRouter([
-  {
-    path: '/',
-    element: (
-      <div>
-        <h1>Hello World</h1>
-        <div className="App">Вот тут будет жить ваше приложение :)</div>
-      </div>
-    ),
-  },
-])
-
-function App() {
-  useEffect(() => {
-    const fetchServerData = async () => {
-      const url = `http://localhost:${__SERVER_PORT__}`
-      const response = await fetch(url)
-      const data = await response.json()
-      console.log(data)
-    }
-
-    fetchServerData()
-  }, [])
-  return <RouterProvider router={router} />
-}
-
-export default App
