+++ packages/client/src/entities/user/api/user-api.ts
@@ -1,3 +1,38 @@
-export const example = () => {
-  console.log('user-api')
+const BASE_URL = 'https://ya-praktikum.tech/api/v2'
+
+const login = (data: string) => {
+  return fetch(`${BASE_URL}/auth/signin`, {
+    method: 'POST',
+    mode: 'cors',
+    credentials: 'include',
+    headers: {
+      'Content-Type': 'application/json',
+    },
+    body: data,
+  })
 }
+
+const logout = () => {
+  return fetch(`${BASE_URL}/auth/logout`, {
+    method: 'POST',
+    mode: 'cors',
+    credentials: 'include',
+    headers: {
+      'Content-Type': 'application/json',
+    },
+  })
+}
+
+const register = (data: string) => {
+  return fetch(`${BASE_URL}/auth/signup`, {
+    method: 'POST',
+    mode: 'cors',
+    credentials: 'include',
+    headers: {
+      'Content-Type': 'application/json',
+    },
+    body: data,
+  })
+}
+
+export { login, logout, register }
