+++ packages/client/src/entities/user/constants/auth.ts
@@ -1,5 +1,21 @@
-// TODO: Переместить в папку entities/user/constants
-const INPUTS = [
+const LOGIN_INPUTS = [
+  {
+    id: 1,
+    selector: 'login',
+    name: 'login',
+    type: 'text',
+    label: 'Логин',
+  },
+  {
+    id: 2,
+    selector: 'password',
+    name: 'password',
+    type: 'password',
+    label: 'Пароль',
+  },
+]
+
+const REGISTER_INPUTS = [
   {
     id: 1,
     name: 'first_name',
@@ -51,4 +67,4 @@ const INPUTS = [
   },
 ]
 
-export { INPUTS }
+export { LOGIN_INPUTS, REGISTER_INPUTS }
