+++ packages/client/src/entities/user/constants/errorOptions.ts
@@ -1,0 +1,11 @@
+const ERROR_OPTIONS = {
+  errorsContainer: '#second-error-field',
+  errorLabelStyle: {
+    fontSize: '0.8rem',
+    lineHeight: '1rem',
+    color: 'rgb(184, 50, 50)',
+  },
+  errorLabelCssClass: 'auth-error-label',
+}
+
+export { ERROR_OPTIONS }
