+++ packages/client/src/entities/user/constants/index.ts
@@ -1,0 +1,2 @@
+export * from './errorOptions'
+export * from './auth'
