--- packages/client/src/pages/login/model/constants.ts
@@ -1,20 +1,0 @@
-// TODO: Переместить в папку entities/user/constants
-
-const INPUTS = [
-  {
-    id: 1,
-    selector: 'login',
-    name: 'login',
-    type: 'text',
-    label: 'Логин',
-  },
-  {
-    id: 2,
-    selector: 'password',
-    name: 'password',
-    type: 'password',
-    label: 'Пароль',
-  },
-]
-
-export { INPUTS }
